requirements.txt
boto==2.45.0
bz2file==0.98
cycler==0.10.0
Cython==0.25.2
decorator==4.0.10
Django==1.10.4
djangorestframework==3.5.3
gensim==0.13.4
html5lib==0.999999999
isodate==0.5.4
matplotlib==1.5.3
networkx==1.11
nltk==3.2.1
numpy==1.11.3
pymongo==3.4.0
pyparsing==2.1.10
pysolr==3.6.0
python-dateutil==2.6.0
python-rake==1.0.7
pytz==2016.10
rdflib==4.2.1
requests==2.12.4
scipy==0.18.1
six==1.10.0
smart-open==1.3.5
SPARQLWrapper==1.8.0
watson-developer-cloud==0.22.0
webencodings==0.5
word2vec==0.9.1
